♻️  made html generator method an util function.
